made HDF5 reader return coordinate arrays matching user-specified precision

+++ vis/python/athena_read.py
@@ -615,7 +615,7 @@ def athdf(filename, raw=False, data=None, quantities=None, dtype=np.float32, lev
             if nx == 1:
                 xm = (x1m, x2m, x3m)[d-1]
                 xp = (x1p, x2p, x3p)[d-1]
-                data[xf] = np.array([xm, xp])
+                data[xf] = np.array([xm, xp], dtype=dtype)
             else:
                 xmin = f.attrs['RootGridX' + repr(d)][0]
                 xmax = f.attrs['RootGridX' + repr(d)][1]
@@ -629,7 +629,7 @@ def athdf(filename, raw=False, data=None, quantities=None, dtype=np.float32, lev
                     data[xf] = face_func(xmin, xmax, xrat_root, nx+1)
                 elif xrat_root == 1.0:
                     if np.all(levels == level):
-                        data[xf] = np.empty(nx + 1)
+                        data[xf] = np.empty(nx + 1, dtype=dtype)
                         for n_block in range(int((nx - 2*num_ghost)
                                                  / (block_size[d-1] - 2*num_ghost))):
                             sample_block = np.where(logical_locations[:, d-1]
@@ -641,15 +641,15 @@ def athdf(filename, raw=False, data=None, quantities=None, dtype=np.float32, lev
                         if num_ghost > 0:
                             raise AthenaError('Cannot use ghost zones with different'
                                               + ' refinement levels')
-                        data[xf] = np.linspace(xmin, xmax, nx + 1)
+                        data[xf] = np.linspace(xmin, xmax, nx + 1, dtype=dtype)
                 else:
                     if num_ghost > 0:
                         raise AthenaError('Ghost zones incompatible with non-uniform'
                                           + ' coordinate spacing')
                     xrat = xrat_root ** (1.0 / 2**level)
-                    data[xf] = (xmin + (1.0-xrat**np.arange(nx+1))
+                    data[xf] = (xmin + (1.0-xrat**np.arange(nx+1, dtype=dtype))
                                 / (1.0-xrat**nx) * (xmax-xmin))
-            data[xv] = np.empty(nx)
+            data[xv] = np.empty(nx, dtype=dtype)
             for i in range(nx):
                 data[xv][i] = center_func(data[xf][i], data[xf][i+1])
